Implementamos el control de permisos en el API de usuarios

+++ users/api.py
@@ -4,10 +4,13 @@ from rest_framework.generics import get_object_or_404
 from rest_framework.response import Response
 from rest_framework.views import APIView
 
+from users.permissions import UserPermission
 from users.serializers import UserSerializer, UserListSerializer, InsertUserSerializer
 
 
 class UserAPI(APIView):
+    permission_classes = [UserPermission]
+
     def get(self, request):
         users = User.objects.all()
         serializer = UserListSerializer(users, many=True)
@@ -24,19 +27,23 @@ class UserAPI(APIView):
 
 
 class UserDetailAPI(APIView):
+    permission_classes = [UserPermission]
 
     def get(self, request, pk):
         user = get_object_or_404(User, pk=pk)
+        self.check_object_permissions(request, user)
         serializer = UserSerializer(user)
         return Response(serializer.data)
 
     def delete(self, request, pk):
         user = get_object_or_404(User, pk=pk)
+        self.check_object_permissions(request, user)
         user.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
 
     def put(self, request, pk):
         user = get_object_or_404(User, pk=pk)
+        self.check_object_permissions(request, user)
         serializer = InsertUserSerializer(user, data=request.data)
         if serializer.is_valid():
             updated_user = serializer.save()
